+++ app/src/org/commcare/services/CommCareFirebaseMessagingService.java
@@ -13,11 +13,9 @@ import com.google.firebase.messaging.RemoteMessage;
 import org.commcare.CommCareNoficationManager;
 import org.commcare.activities.DispatchActivity;
 import org.commcare.dalvik.R;
-import org.commcare.sync.FirebaseMessagingDataSyncer;
 import org.commcare.util.LogTypes;
 import org.commcare.utils.FirebaseMessagingUtil;
 import org.javarosa.core.services.Logger;
-import org.joda.time.DateTime;
 
 import java.util.Map;
 
@@ -29,7 +27,7 @@ import java.util.Map;
 public class CommCareFirebaseMessagingService extends FirebaseMessagingService {
 
     private final static int FCM_NOTIFICATION = R.string.fcm_notification;
-    private enum ActionTypes{
+    enum ActionTypes{
         SYNC,
         INVALID
     }
@@ -60,7 +58,7 @@ public class CommCareFirebaseMessagingService extends FirebaseMessagingService {
 
         FCMMessageData fcmMessageData = new FCMMessageData(payloadData);
 
-        switch(fcmMessageData.action){
+        switch(fcmMessageData.getAction()){
             case SYNC -> {} // trigger sync for fcmMessageData
             default ->
                     Logger.log(LogTypes.TYPE_FCM, "Invalid FCM action");
@@ -105,57 +103,5 @@ public class CommCareFirebaseMessagingService extends FirebaseMessagingService {
                 .setWhen(System.currentTimeMillis());
 
         mNM.notify(FCM_NOTIFICATION, fcmNotification.build());
-    }
-
-    /**
-     * This class is to facilitate handling the FCM Message Data object. It should contain all the
-     * necessary checks and transformations
-     */
-    public class FCMMessageData {
-        private ActionTypes action;
-        private String username;
-        private String domain;
-        private DateTime creationTime;
-
-        private FCMMessageData(Map<String, String> payloadData){
-            this.action = getActionType(payloadData.get("action"));
-            this.username = payloadData.get("username");
-            this.domain = payloadData.get("domain");
-            this.creationTime = convertISO8601ToDateTime(payloadData.get("created_at"));
-        }
-
-        public String getUsername() {
-            return username;
-        }
-
-        public String getDomain() {
-            return domain;
-        }
-
-        public DateTime getCreationDate() {
-            return creationTime;
-        }
-
-        private DateTime convertISO8601ToDateTime(String timeInISO8601) {
-            if (timeInISO8601 == null){
-                return new DateTime();
-            }
-            return new DateTime(timeInISO8601);
-        }
-
-        private ActionTypes getActionType(String action) {
-            if (action == null) {
-                return ActionTypes.INVALID;
-            }
-
-            switch (action.toUpperCase()) {
-                case "SYNC" -> {
-                    return ActionTypes.SYNC;
-                }
-                default -> {
-                    return ActionTypes.INVALID;
-                }
-            }
-        }
     }
 }
